requirements_enhanced.txt
# requirements_enhanced.txt
# Enhanced LiveKit Voice AI Agent with Multi-Agent Architecture

# Core LiveKit Dependencies (1.0+)
livekit-agents[openai,silero,deepgram,elevenlabs,cartesia,turn-detector]
livekit-api
livekit-protocol

# Enhanced TTS Options
livekit-plugins-elevenlabs
livekit-plugins-cartesia
livekit-plugins-openai

# Enhanced STT Options
livekit-plugins-deepgram
livekit-plugins-google

# Core AI Models
livekit-plugins-silero
livekit-plugins-turn-detector

# Qdrant Vector Database
qdrant-client
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-openai

# AI and Embeddings
openai
sentence-transformers
fastembed

# Essential Dependencies
python-dotenv
pydantic
pydantic-settings
asyncio-extras
aiofiles

# Document Processing (Enhanced)
pypdf2
python-docx
markdown
pandas
openpyxl
xlrd

# Database and Memory

redis


# Performance and Monitoring
numpy
scipy
psutil


# Natural Language Processing
nltk
spacy
textblob

# HTTP and API
httpx
requests
websockets

# Async and Concurrency

asyncio-throttle

# Logging and Debugging
colorlog
structlog
rich

# Configuration and Validation
python-decouple
marshmallow

# Testing (Development)
pytest
pytest-asyncio
pytest-mock

# Security
cryptography
python-jose

# Optional: Advanced Features
# Uncomment if you want these features:

# Advanced Speech Processing
# pyaudio>=0.2.11
# librosa>=0.10.0
# soundfile>=0.12.0

# Machine Learning
# torch>=2.1.0
# transformers>=4.35.0
# accelerate>=0.24.0

# Distributed Computing
# celery>=5.3.0
# flower>=2.0.0

# Advanced Monitoring
# grafana-client>=3.7.0
# influxdb-client>=1.38.0

# Cloud Storage
# boto3>=1.34.0  # AWS
# azure-storage-blob>=12.19.0  # Azure
# google-cloud-storage>=2.10.0  # GCP

# Production WSGI/ASGI
# gunicorn>=21.2.0
# uvicorn>=0.24.0

# Memory Optimization
# pympler>=0.9
# memory-profiler>=0.61.0